requirements.txt
MetaTrader5>=5.0.45
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
pytz>=2023.3
feedparser>=6.0.10
